refactor(web): rename main PaginedTable filename
