+++ apps/web/src/pages/blocks.tsx
@@ -6,8 +6,8 @@ import { getPaginationParams } from "~/utils/pagination";
 import { BlobGasUsageDisplay } from "~/components/Displays/BlobGasUsageDisplay";
 import { EtherUnitDisplay } from "~/components/Displays/EtherUnitDisplay";
 import { Link } from "~/components/Link";
-import type { PaginatedTableQueryFilters } from "~/components/PaginatedTable/PaginatedTable";
-import { PaginatedTable } from "~/components/PaginatedTable/PaginatedTable";
+import type { PaginatedTableQueryFilters } from "~/components/PaginatedTable";
+import { PaginatedTable } from "~/components/PaginatedTable";
 import { Table } from "~/components/Table";
 import { api } from "~/api-client";
 import NextError from "~/pages/_error";
