+++ apps/web/src/pages/txs.tsx
@@ -5,8 +5,8 @@ import { useRouter } from "next/router";
 import { getPaginationParams } from "~/utils/pagination";
 import { EtherUnitDisplay } from "~/components/Displays/EtherUnitDisplay";
 import { Link } from "~/components/Link";
-import type { PaginatedTableQueryFilters } from "~/components/PaginatedTable/PaginatedTable";
-import { PaginatedTable } from "~/components/PaginatedTable/PaginatedTable";
+import type { PaginatedTableQueryFilters } from "~/components/PaginatedTable";
+import { PaginatedTable } from "~/components/PaginatedTable";
 import { RollupIcon } from "~/components/RollupIcon";
 import { Table } from "~/components/Table";
 import { api } from "~/api-client";
